requirements.txt
streamlit
requests
google-cloud-dialogflow
google-cloud-translate
google-cloud-texttospeech
google-cloud-speech
streamlit-webrtc
av
numpy
google-cloud-speech
python-dotenv==1.0.1
PyJWT==2.8.0
google-auth
